style: cleaned out code, refactor: design new modal
Design a new modal for pack Activities with offset that utilizes Aerie's Modal to match with design changes. Cleaned out unnecessary code.
