+++ src/components/ui/DataGrid/BulkActionDataGrid.svelte
@@ -21,8 +21,8 @@
   import type { PermissionCheck } from '../../../types/permissions';
   import { isDeleteEvent } from '../../../utilities/keyboardEvents';
   import { permissionHandler } from '../../../utilities/permissionHandler';
+  import PackActivitiesOffsetModal from '../../modals/PackActivitiesOffsetModal.svelte';
   import DataGrid from '../../ui/DataGrid/DataGrid.svelte';
-  import PackActivitiesOffsetSelect from '../PackActivitesOffsetSelect.svelte';
 
   export let autoSizeColumnsToFit: boolean = true;
   export let columnDefs: ColDef[];
@@ -273,5 +273,5 @@
 </DataGrid>
 
 {#if showPackOffsetDialog}
-  <PackActivitiesOffsetSelect on:cancel={() => (showPackOffsetDialog = false)} on:pack={bulkPackItemsWithOffset} />
+  <PackActivitiesOffsetModal on:cancel={() => (showPackOffsetDialog = false)} on:pack={bulkPackItemsWithOffset} />
 {/if}
